moved createGetStaticPaths to its own file and made it generic
