+++ src/photos/PhotoPage.tsx
@@ -28,53 +28,3 @@ export const PhotoPage: FC<Props> = ({ folderName, currentPhoto }: Props) => {
         </>
     )
 }
-
-// export const createGetStaticProps =
-//     (folderName: string): GetStaticProps =>
-//     async (context) => {
-//         const results = await getResults(folderName)
-//         const reducedResults: ImageProps[] = []
-//         let i = 0
-
-//         for (const result of results.resources) {
-//             reducedResults.push({
-//                 id: i,
-//                 height: String(result.height),
-//                 width: String(result.width),
-//                 public_id: result.public_id,
-//                 format: result.format,
-//             })
-//             i++
-//         }
-
-//         const currentPhoto = reducedResults.find((img) => img.id === Number(context.params?.photoId))
-//         if (currentPhoto) {
-//             currentPhoto.blurDataUrl = await getBase64ImageUrl(currentPhoto)
-//         }
-
-//         return {
-//             props: {
-//                 currentPhoto,
-//             },
-//         }
-//     }
-
-// export function createGetStaticPaths(folderName: string) {
-//     return async function getStaticPaths() {
-//         const results = await cloudinary.v2.search
-//             .expression(`folder:${folderName}/*`)
-//             .sort_by('public_id', 'desc')
-//             .max_results(400)
-//             .execute()
-
-//         const fullPaths = []
-//         for (let i = 0; i < results.resources.length; i++) {
-//             fullPaths.push({ params: { photoId: i.toString() } })
-//         }
-
-//         return {
-//             paths: fullPaths,
-//             fallback: false,
-//         }
-//     }
-// }
